requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
et-xmlfile==1.1.0
Flask==2.1.2
Flask-Login==0.6.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.22.3
openpyxl==3.0.9
pandas==1.4.2
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
requests==2.27.1
six==1.16.0
spotipy==2.19.0
SQLAlchemy==1.4.36
urllib3==1.26.9
Werkzeug==2.1.2
zipp==3.8.0
